Reimplemented simplicial_laplacian and added simplicial_visualization
