+++ python/topology/__init__.py
@@ -13,6 +13,6 @@ __copyright__ = "Copyright 2023, 2025  All rights reserved."
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-class TopologyException(Exception):
+class TopologyException(BaseException):
     def __init__(self, *args, **kwargs):
         super(TopologyException, self).__init__(args, kwargs)
